+++ R/access.R
@@ -18,6 +18,8 @@ access_types <- c("exists" = 0L, "read" = 4L, "write" = 2L, "execute" = 1L)
 #'
 #' file_exists("WOMBATS")
 file_access <- function(path, mode = "exists") {
+  assert_no_missing(path)
+
   path <- path_expand(path)
   mode <- match.arg(mode, names(access_types), several.ok = TRUE)
   mode <- sum(access_types[mode])
@@ -28,6 +30,7 @@ file_access <- function(path, mode = "exists") {
 #' @rdname file_access
 #' @export
 file_exists <- function(path) {
+  path <- path_expand(path)
   res <- file_info(path)
   setNames(!is.na(res$type), res$path)
 }
@@ -35,6 +38,7 @@ file_exists <- function(path) {
 #' @rdname file_access
 #' @export
 dir_exists <- function(path) {
+  path <- path_expand(path)
   res <- is_dir(path)
   !is.na(res) & res
 }
@@ -42,6 +46,7 @@ dir_exists <- function(path) {
 #' @rdname file_access
 #' @export
 link_exists <- function(path) {
+  path <- path_expand(path)
   res <- is_link(path)
   !is.na(res) & res
 }