+++ R/file.R
@@ -85,10 +85,12 @@ file_types <- c(
 #' file_chmod(x, "u+wr")
 #' file_info(x)$permissions
 file_chmod <- function(path, mode) {
-  stopifnot(length(mode) == 1)
-  mode <- as_fs_perms(mode, mode = file_info(path)$permissions)
+  assert_no_missing(path)
 
   path <- path_expand(path)
+
+  stopifnot(length(mode) == 1)
+  mode <- as_fs_perms(mode, mode = file_info(path)$permissions)
 
   chmod_(path, mode)
 
@@ -103,6 +105,8 @@ file_chmod <- function(path, mode) {
 #'   name.
 #' @export
 file_chown <- function(path, user_id = NULL, group_id = NULL) {
+  assert_no_missing(path)
+
   path <- path_expand(path)
 
   if (is.null(user_id)) {
@@ -137,9 +141,52 @@ file_chown <- function(path, user_id = NULL, group_id = NULL) {
 #' @inheritParams utils::browseURL
 #' @export
 file_show <- function(path = ".", browser = getOption("browser")) {
+  assert_no_missing(path)
+
+  path <- path_expand(path)
+
   for (p in path) {
     browseURL(p)
   }
 
   invisible(path_tidy(path))
+}
+
+
+#' Move or rename files
+#'
+#' Compared to [file.rename] `file_move()` always fails if it is unable to move
+#' a file, rather than signaling a Warning and returning an error code.
+#' @template fs
+#' @param new_path New file path. If `new_path` is existing directory, the file
+#'   will be moved into that directory; otherwise it will be moved/renamed to
+#'   the full path.
+#'
+#'   Should either be the same length as `path`, or a single directory.
+#' @return The new path (invisibly).
+#' @examples
+#' file_create("foo")
+#' file_move("foo", "bar")
+#' file_exists(c("foo", "bar"))
+#' file_delete("bar")
+#' @export
+file_move <- function(path, new_path) {
+  assert_no_missing(path)
+  assert_no_missing(new_path)
+
+  path <- path_expand(path)
+  new_path <- path_expand(new_path)
+
+  is_directory <- file_exists(new_path) && is_dir(new_path)
+
+  if (length(new_path) == 1 && is_directory[[1]]) {
+    new_path <- rep(new_path, length(path))
+  }
+  stopifnot(length(path) == length(new_path))
+
+  new_path[is_directory] <- path(new_path[is_directory], basename(path))
+
+  move_(path, new_path)
+
+  invisible(path_tidy(new_path))
 }