+++ R/list.R
@@ -28,6 +28,10 @@
 #' dir_ls("base", recursive = TRUE, glob = "*.R")
 dir_ls <- function(path = ".", all = FALSE, recursive = FALSE, type = "any",
                    glob = NULL, regexp = NULL, invert = FALSE, ...) {
+  assert_no_missing(path)
+
+  path <- path_expand(path)
+
   files <- as.character(dir_map(path, identity, all, recursive, type))
 
   path_filter(files, glob, regexp, invert = invert,...)
@@ -50,9 +54,11 @@ directory_entry_types <- c(
 #' dir_map("base", identity)
 #' @export
 dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
-  type <- match.arg(type, names(directory_entry_types), several.ok = TRUE)
+  assert_no_missing(path)
 
   path <- path_expand(path)
+
+  type <- match.arg(type, names(directory_entry_types), several.ok = TRUE)
 
   dir_map_(path, fun, all, sum(directory_entry_types[type]), recursive)
 }
@@ -62,6 +68,10 @@ dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any
 #' dir_walk("base", str)
 #' @export
 dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any") {
+  assert_no_missing(path)
+
+  path <- path_expand(path)
+
   dir_map(path, fun, all, recursive, type)
   invisible(path_tidy(path))
 }
@@ -75,7 +85,10 @@ dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "an
 #' link_delete("base")
 dir_info <- function(path = ".", all = FALSE, recursive = FALSE,
                      type = "any", regexp = NULL, glob = NULL, ...) {
+  assert_no_missing(path)
+
+  path <- path_expand(path)
+
   file_info(dir_ls(path = path, all = all, recursive = recursive, type = type,
     regexp = regexp, glob = glob, ...))
 }
-