+++ R/temp.R
@@ -31,12 +31,20 @@ env$temp_names <- character()
 #' while (!is.null(file_temp_pop())) next
 #' file_temp_pop()
 file_temp <- function(pattern = "file", tmp_dir = tempdir(), ext = "") {
+  assert_no_missing(tmp_dir)
+
+  tmp_dir <- path_expand(tmp_dir)
+
   path_tidy(file_temp_pop() %||% tempfile(pattern, tmp_dir, ext))
 }
 
 #' @export
 #' @rdname file_temp
 file_temp_push <- function(path) {
+  assert_no_missing(path)
+
+  path <- path_expand(path)
+
   env$temp_names <- c(env$temp_names, path)
 
   invisible(path_tidy(path))