+++ src/path.cc
@@ -29,6 +29,9 @@ CharacterVector path_(List paths, const char* ext) {
   char* b = buf;
   for (R_xlen_t c = 0; c < max_col; ++c) {
     R_xlen_t len = Rf_xlength(VECTOR_ELT(paths, c));
+    if (len == 0) {
+      return CharacterVector();
+    }
     if (len > max_row) {
       max_row = len;
     }
@@ -46,7 +49,7 @@ CharacterVector path_(List paths, const char* ext) {
           break;
         }
 
-        const char* s = Rf_translateCharUTF8(str);
+        const char* s = CHAR(str);
         strcpy(b, s);
         b += strlen(s);
 
@@ -66,7 +69,22 @@ CharacterVector path_(List paths, const char* ext) {
         b += strlen(ext) + 1;
       }
       *b = '\0';
-      out[r] = buf;
+      out[r] = Rf_mkCharCE(buf, CE_UTF8);
+    }
+  }
+  return out;
+}
+
+// [[Rcpp::export]]
+CharacterVector expand_(CharacterVector path) {
+  CharacterVector out = CharacterVector(path.size());
+
+  for (R_xlen_t i = 0; i < Rf_xlength(out); ++i) {
+    if (STRING_ELT(path, i) == R_NaString) {
+      SET_STRING_ELT(out, i, R_NaString);
+    } else {
+      const char* p = CHAR(STRING_ELT(path, i));
+      SET_STRING_ELT(out, i, Rf_mkCharCE(R_ExpandFileName(p), CE_UTF8));
     }
   }
   return out;