+++ tests/testthat/test-file.R
@@ -1,16 +1,18 @@
 context("test-file.R")
 
 describe("file_info", {
-  with_dir_tree(list("foo/bar" = "test"), {
+  with_dir_tree(list(
+      "foo/bar" = "test",
+      "NA" = ""), {
     link_create(path_abs("foo"), "foo2")
 
     it("returns a correct tibble", {
-      x <- file_info(c("foo", "foo/bar", "foo2"))
+      x <- file_info(c("foo", "foo/bar", "foo2", "NA"))
       expect_is(x, c("tbl", "tbl_df", "data.frame"))
       expect_length(x, 18)
-      expect_equal(nrow(x), 3)
-      expect_equal(as.character(x$path), c("foo", "foo/bar", "foo2"))
-      expect_equal(as.character(x$type), c("directory", "file", "symlink"))
+      expect_equal(nrow(x), 4)
+      expect_equal(as.character(x$path), c("foo", "foo/bar", "foo2", "NA"))
+      expect_equal(as.character(x$type), c("directory", "file", "symlink", "file"))
     })
 
     it("returns NA if a file does not exist", {
@@ -20,6 +22,14 @@ describe("file_info", {
       expect_equal(nrow(x), 1)
       expect_equal(as.character(x$path), "missing")
       expect_equal(sum(is.na(x)), 17)
+    })
+    it("returns NA on NA input", {
+      x <- file_info(NA_character_)
+      expect_is(x, c("tbl", "tbl_df", "data.frame"))
+      expect_length(x, 18)
+      expect_equal(nrow(x), 1)
+      expect_equal(as.character(x$path), NA_character_)
+      expect_equal(sum(is.na(x)), 18)
     })
   })
 })
@@ -65,21 +75,14 @@ if (!is_windows()) {
       it("errors if given an invalid mode", {
         expect_error(file_chmod("foo", "g+S"), "Invalid mode 'g\\+S'")
       })
+      it("errors on missing input", {
+        expect_error(file_chmod(NA, "u+x"), class = "invalid_argument")
+      })
     })
   })
 }
 
-describe("file_delete", {
-  with_dir_tree(list("foo/bar" = "test"), {
-    it("returns the input path and deletes the file", {
-      expect_true(file_exists("foo/bar"))
-      expect_equal(file_delete("foo/bar"), "foo/bar")
-      expect_false(file_exists("foo/bar"))
-    })
-  })
-})
-
-describe("file_delete", {
+describe("file_copy", {
   with_dir_tree(list("foo/bar" = "test"), {
     it("returns the new path and copies the file", {
       expect_true(file_exists("foo/bar"))
@@ -93,13 +96,17 @@ describe("file_delete", {
       expect_error(file_copy("baz/bar", "foo/bar3"), class = "ENOENT")
       expect_error(file_copy("foo/bar", "baz/qux"), class = "ENOENT")
     })
+    it("errors on missing input", {
+      expect_error(file_copy(NA, "foo"), class = "invalid_argument")
+      expect_error(file_copy("foo/bar", NA), class = "invalid_argument")
+    })
   })
 })
 
 describe("file_chown", {
+  skip("need elevated permissions to change uid")
   with_dir_tree(list("foo/bar" = "test"), {
     it("changes the ownership of a file, returns the input path", {
-      skip("need elevated permissions to change uid")
 
       # Make everyone have write access, so we can delete this after changing ownership
       file_chmod("foo/bar", "a+w")
@@ -108,5 +115,30 @@ describe("file_chown", {
       expect_equal(file_chown("foo/bar", user_id = 0), "foo/bar")
       expect_true(file_info("foo/bar")$user_id == 0)
     })
+    it("errors on missing input", {
+      expect_error(file_copy(NA, user_id = 0), class = "invalid_argument")
+    })
+  })
+})
+
+describe("file_move", {
+  it("renames files in the same directory", {
+    with_dir_tree(list("foo" = "test"), {
+      expect_true(file_exists("foo"))
+      expect_equal(file_move("foo", "bar"), "bar")
+      expect_false(file_exists("foo"))
+      expect_true(file_exists("bar"))
+    })
+  })
+
+  it("renames files in different directories", {
+    with_dir_tree(list("foo/bar" = "test", "foo2"), {
+      expect_true(file_exists("foo/bar"))
+      expect_true(file_exists("foo2"))
+    })
+  })
+  it("errors on missing input", {
+    expect_error(file_move(NA, "foo2"), class = "invalid_argument")
+    expect_error(file_move("foo", NA), class = "invalid_argument")
   })
 })